Separating server API from LLM related functions
Moving utility functions into separate module.
